fix wording in tests and move tests to selectors
